\name{chanames}
\alias{chanames}
\alias{varnames}
\title{
coda package related generic functions
}
\description{
\pkg{coda} package related generic functions.
}
\usage{
chanames(x, ...)
varnames(x, ...)
}
\arguments{
  \item{x}{
MCMC object.
}
  \item{\dots}{
Other arguments.
}
}
\value{
See corrsponding help pages.
}
\author{
Peter Solymos
}
\seealso{
\code{\link[coda]{chanames}}
\code{\link[coda]{varnames}}
}
\keyword{utils}
