\name{crosscorr.plot}
\alias{crosscorr.plot}
\title{
Generic after similar coda function
}
\description{
Generic after similar coda function
}
\usage{
crosscorr.plot(x, ...)
}
\arguments{
  \item{x}{
MCMC object.
}
  \item{\dots}{
Other arguments.
}
}
\value{
See corresponding help page
}
\author{
Peter Solymos
}
\seealso{
\code{\link[coda]{crosscorr.plot}}
}
\keyword{utils}

