\name{cumuplot}
\alias{cumuplot}
\title{
Generic after similar coda function
}
\description{
Generic after similar coda function
}
\usage{
cumuplot(x, ...)
}
\arguments{
  \item{x}{
MCMC object.
}
  \item{\dots}{
Other arguments.
}
}
\value{
See corresponding help page
}
\author{
Peter Solymos
}
\seealso{
\code{\link[coda]{cumuplot}}
}
\keyword{utils}

