\name{custommodel-class}
\Rdversion{1.1}
\docType{class}
\alias{custommodel-class}

\title{Class \code{"custommodel"}}
\description{
Stands for the 'custommodel' S3 class from \pkg{dclone} package.
}
\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}
\section{Extends}{
Class \code{"\linkS4class{dcModel}"}, directly.
}
\section{Methods}{
No methods defined with class "custommodel" in the signature.
}
\author{
Peter Solymos
}
\seealso{
\code{\link[dclone]{custommodel}}
}
\examples{
showClass("custommodel")
}
\keyword{classes}
