\name{dcArgs-class}
\Rdversion{1.1}
\docType{class}
\alias{dcArgs-class}

\title{Class \code{"dcArgs"}}
\description{
A class union for \code{NULL} and \code{"character"}.
}
\section{Objects from the Class}{A virtual Class: 
No objects may be created from it.}
\section{Methods}{
No methods defined with class "dcArgs" in the signature.
}
\author{
Peter Solymos
}
\examples{
showClass("dcArgs")
}
\keyword{classes}
