\name{dcdiag-class}
\Rdversion{1.1}
\docType{class}
\alias{dcdiag-class}

\title{Class \code{"dcdiag"}}
\description{
Stands for the 'dcdiag' S3 class from \pkg{dclone} package.
}
\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}
\section{Extends}{
Class \code{"\linkS4class{dcDiag}"}, directly.
}
\section{Methods}{
No methods defined with class "dcdiag" in the signature.
}
\author{
Peter Solymos
}
\seealso{
\code{\link[dclone]{dcdiag}}
}
\examples{
showClass("dcdiag")
}
\keyword{classes}
