\name{dcDiag-class}
\Rdversion{1.1}
\docType{class}
\alias{dcDiag-class}

\title{Class \code{"dcDiag"}}
\description{
Virtual class for data cloning convergence diagnostics.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "dcDiag" in the signature.
}
\author{
Peter Solymos
}
\seealso{
\code{\link[dclone]{dcdiag}}
}
\examples{
showClass("dcDiag")
}
\keyword{classes}
