\name{dcFunction-class}
\Rdversion{1.1}
\docType{class}
\alias{dcFunction-class}

\title{Class \code{"dcFunction"}}
\description{
Virtual class for BUGS/JAGS models defined as functions.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "dcFunction" in the signature.
}
\author{
Peter Solymos
}
\examples{
showClass("dcFunction")
}
\keyword{classes}
