\name{dcInits-class}
\Rdversion{1.1}
\docType{class}
\alias{dcInits-class}

\title{Class \code{"dcInits"}}
\description{
Virtual class for initial values.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "dcInits" in the signature.
}
\author{
Peter Solymos
}
\examples{
showClass("dcInits")
}
\keyword{classes}
