\name{dcModel-class}
\Rdversion{1.1}
\docType{class}
\alias{dcModel-class}

\title{Class \code{"dcModel"}}
\description{
Virtual class for BUGS/JAGS models.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "dcModel" in the signature.
}
\author{
Peter Solymos
}
\examples{
showClass("dcModel")
}
\keyword{classes}
