\name{dcParams-class}
\Rdversion{1.1}
\docType{class}
\alias{dcParams-class}

\title{Class \code{"dcParams"}}
\description{
Virtual class for model parameters to monitor.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "dcParams" in the signature.
}
\author{
Peter Solymos
}
\examples{
showClass("dcParams")
}
\keyword{classes}
