\name{dctable-class}
\Rdversion{1.1}
\docType{class}
\alias{dctable-class}

\title{Class \code{"dctable"}}
\description{
Stands for the 'dctable' S3 class from \pkg{dclone} package.
}
\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}
\section{Extends}{
Class \code{"\linkS4class{dcTable}"}, directly.
}
\section{Methods}{
No methods defined with class "dctable" in the signature.
}
\author{
Peter Solymos
}
\seealso{
\code{\link[dclone]{dctable}}
}
\examples{
showClass("dctable")
}
\keyword{classes}
