\name{dcTable-class}
\Rdversion{1.1}
\docType{class}
\alias{dcTable-class}

\title{Class \code{"dcTable"}}
\description{
Posterior statistics from iterative fit, virtual class.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "dcTable" in the signature.
}
\author{
Peter Solymos
}
\examples{
showClass("dcTable")
}
\keyword{classes}
