\name{mcmc-class}
\Rdversion{1.1}
\docType{class}
\alias{mcmc-class}

\title{Class \code{"mcmc"}}
\description{
Stands for the 'mcmc' S3 class from \pkg{coda} package.
}
\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}
\section{Extends}{
Class \code{"\linkS4class{MCMClist}"}, directly.
}
\section{Methods}{
No methods defined with class "mcmc" in the signature.
}
\author{
Peter Solymos
}
\seealso{
\code{\link[coda]{mcmc}}
}
\examples{
showClass("mcmc")
}
\keyword{classes}
