\name{mcmc.list-class}
\Rdversion{1.1}
\docType{class}
\alias{mcmc.list-class}

\title{Class \code{"mcmc.list"}}
\description{
Stands for the 'mcmc.list' S3 class from \pkg{coda} package.
}
\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}
\section{Extends}{
Class \code{"\linkS4class{MCMClist}"}, directly.
}
\section{Methods}{
No methods defined with class "mcmc.list" in the signature.
}
\author{
Peter Solymos
}
\seealso{
\code{\link[coda]{mcmc.list}}
}
\examples{
showClass("mcmc.list")
}
\keyword{classes}
