\name{mcmc.list.dc-class}
\Rdversion{1.1}
\docType{class}
\alias{mcmc.list.dc-class}

\title{Class \code{"mcmc.list.dc"}}
\description{
Stands for the 'mcmc.list.dc' S3 class from \pkg{dclone} package.
}
\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}
\section{Extends}{
Class \code{"\linkS4class{MCMClist}"}, directly.
}
\section{Methods}{
No methods defined with class "mcmc.list.dc" in the signature.
}
\author{
Peter Solymos
}
\seealso{
\code{\link[coda]{mcmc.list}}, \code{\link[dclone]{jags.fit}}
}
\examples{
showClass("mcmc.list.dc")
}
\keyword{classes}
