\name{nClones-class}
\Rdversion{1.1}
\docType{class}
\alias{nClones-class}

\title{Class \code{"nClones"}}
\description{
Number of clones, virtual class.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "nClones" in the signature.
}
\author{
Peter Solymos
}
\examples{
showClass("nClones")
}
\keyword{classes}
